+++ core/templates/core/base.html
@@ -154,7 +154,7 @@
 
         {% block footer %}
             <footer class="text-center">
-                <a href='https://github.com/Egorka96/med-org-portal' target='_blank'>&copy; Egor Razzhivkin 2021</a>
+                <a href='https://github.com/Egorka96/med-org-portal' target='_blank'>&copy; Egor Razzhivkin 2020 - 2021</a>
             </footer>
         {% endblock %}
