+++ core/templates/core/forgot_password.html
@@ -9,7 +9,7 @@
             {% csrf_token %}
             <form method="post">
                 {% csrf_token %}
-                <h5>Введите свою электронную почту, которая привязана к аккаунту.</h5>
+                <h5>Введите адрес электронной почту, связанного с вашим аккаунтов в личном кабинете.</h5>
                 {% bootstrap_form form %}
                 <button type="submit" class="btn btn-primary btn-block btn-login">Отправить</button>
             </form>
