+++ project/settings.py
@@ -166,9 +166,10 @@ EMAIL_USER_CREDENTIALS_TEXT = """
 """
 
 EMAIL_FORGOT_USER_TEXT = """
-Вам была создана учетная запись в личном кабинете медцентра "{{ med_center_name }}".
-Адрес личного кабинета - {{ portal_url }}.
-Пароль - {{password}}
+    Ваш email адрес был указан для сброса текущего пароля в личном кабинете {{ med_center_name }}".
+    Адрес личного кабинета - {{ portal_url }}.
+    Ваш новый временный пароль - новый_пароль. - {{password}}.
+    При первом входе в личный кабинет, система попросит вас придумать свой новый пароль.
 """
 
 try:
